refactor: Mutualize viewer refactor: Split photos-model into model and timeline
