+++ testcafe/tests/drive/public-viewer-feature.js
@@ -151,21 +151,21 @@ test(`${TEST_PUBLIC_VIEWER_ZIP}`, async t => {
     width: 140,
     y: 180
   })
-
   await t.fixtureCtx.vr.takeScreenshotAndUpload(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_ZIP}-wholeFolder`,
     true
   )
+
   await publicViewerPage.checkViewerNavigation(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_ZIP}-nav`,
     data.FILE_ZIP,
     3
   )
 
-  await publicViewerPage.checkCommonViewerDownload(data.FILE_ZIP)
+  await publicViewerPage.openFileAndCheckCommonViewerDownload(data.FILE_ZIP)
   t.ctx.fileDownloaded = data.FILE_ZIP
 
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_ZIP}-1`,
     data.FILE_ZIP
   )
@@ -185,7 +185,7 @@ test(`${TEST_PUBLIC_VIEWER_ZIP}`, async t => {
   })
   t.ctx.fileDownloaded = data.FILE_PPTX
 
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_PPTX}-1`,
     data.FILE_PPTX
   )
@@ -198,10 +198,10 @@ test(`${TEST_PUBLIC_VIEWER_ZIP}`, async t => {
 
 test(`${TEST_PUBLIC_VIEWER_IMG}`, async t => {
   console.group(`↳ ℹ️  ${FEATURE_PREFIX} : ${TEST_PUBLIC_VIEWER_IMG}`)
-  await publicViewerPage.checkCommonViewerDownload(data.FILE_IMG)
+  await publicViewerPage.openFileAndCheckCommonViewerDownload(data.FILE_IMG)
   t.ctx.fileDownloaded = data.FILE_IMG
 
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_IMG}-1`,
     data.FILE_IMG,
     'img'
@@ -215,7 +215,7 @@ test(`${TEST_PUBLIC_VIEWER_IMG}`, async t => {
 
 test(`${TEST_PUBLIC_VIEWER_AUDIO}`, async t => {
   console.group(`↳ ℹ️  ${FEATURE_PREFIX} : ${TEST_PUBLIC_VIEWER_AUDIO}`)
-  await publicViewerPage.checkCommonViewerDownload(data.FILE_AUDIO)
+  await publicViewerPage.openFileAndCheckCommonViewerDownload(data.FILE_AUDIO)
   t.ctx.fileDownloaded = data.FILE_AUDIO
 
   //mask on loading bar
@@ -225,7 +225,7 @@ test(`${TEST_PUBLIC_VIEWER_AUDIO}`, async t => {
     width: 100,
     y: 623
   })
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_AUDIO}-1`,
     data.FILE_AUDIO,
     'audio',
@@ -248,10 +248,10 @@ test(`${TEST_PUBLIC_VIEWER_AUDIO}`, async t => {
 
 test(`${TEST_PUBLIC_VIEWER_VIDEO}`, async t => {
   console.group(`↳ ℹ️  ${FEATURE_PREFIX} : ${TEST_PUBLIC_VIEWER_VIDEO}`)
-  await publicViewerPage.checkCommonViewerDownload(data.FILE_VIDEO)
+  await publicViewerPage.openFileAndCheckCommonViewerDownload(data.FILE_VIDEO)
   t.ctx.fileDownloaded = data.FILE_VIDEO
 
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_VIDEO}-1`,
     data.FILE_VIDEO,
     'video'
@@ -265,10 +265,10 @@ test(`${TEST_PUBLIC_VIEWER_VIDEO}`, async t => {
 
 test(`${TEST_PUBLIC_VIEWER_TXT}`, async t => {
   console.group(`↳ ℹ️  ${FEATURE_PREFIX} : ${TEST_PUBLIC_VIEWER_TXT}`)
-  await publicViewerPage.checkCommonViewerDownload(data.FILE_TXT)
+  await publicViewerPage.openFileAndCheckCommonViewerDownload(data.FILE_TXT)
   t.ctx.fileDownloaded = data.FILE_TXT
 
-  await publicViewerPage.openFileAndcheckCommonViewerDownload(
+  await publicViewerPage.openFileAndCheckPublicViewer(
     `${FEATURE_PREFIX}/${TEST_PUBLIC_VIEWER_TXT}-1`,
     data.FILE_TXT,
     'txt'
