+++ testcafe/tests/drive/viewer-feature.js
@@ -75,7 +75,9 @@ test('Viewer : checking common features for all files (expect PDF)', async t => 
       data.FOLDER_DATE_TIME,
       t.ctx.fileNameListNoPDF[i]
     )
-    await viewerPage.checkCommonViewerDownload(t.ctx.fileNameListNoPDF[i])
+    await viewerPage.openFileAndCheckCommonViewerDownload(
+      t.ctx.fileNameListNoPDF[i]
+    )
     console.groupEnd()
   }
 }).after(async t => {
