+++ testcafe/tests/pages/drive-viewer/drive-viewer-model.js
@@ -1,38 +1,19 @@
-import { t, Selector } from 'testcafe'
+import { t } from 'testcafe'
 import {
   getElementWithTestId,
   isExistingAndVisibile
 } from '../../helpers/utils'
 import DrivePage from '../drive/drive-model'
+import Viewer from '../viewer/viewer-model'
 
 const drivePage = new DrivePage()
 
-export default class Page {
+export default class ViewerDrive extends Viewer {
   constructor() {
-    this.spinner = Selector('[class*="c-spinner"]')
-    this.viewerWrapper = getElementWithTestId('viewer-wrapper')
-    this.viewerControls = getElementWithTestId('pho-viewer-controls')
-    this.viewerToolbar = getElementWithTestId('viewer-toolbar')
-    this.btnDownloadViewerToolbar = getElementWithTestId(
-      'viewer-toolbar-download'
-    )
-
-    // Navigation in viewer
-    this.viewerNavNext = getElementWithTestId('viewer-nav--next')
-    this.viewerNavNextBtn = this.viewerNavNext.find(
-      '[class*="pho-viewer-nav-arrow"]'
-    )
-    this.viewerNavPrevious = getElementWithTestId('viewer-nav--previous')
-    this.viewerNavPreviousBtn = this.viewerNavPrevious.find(
-      '[class*="pho-viewer-nav-arrow"]'
-    )
-    this.viewerBtnClose = getElementWithTestId('btn-viewer-toolbar-close')
-
+    super()
     //Viewer type
     this.audioViewer = getElementWithTestId('viewer-audio')
     this.audioViewerControls = this.audioViewer.find('audio')
-    this.imageViewer = getElementWithTestId('viewer-image')
-    this.imageViewerContent = this.imageViewer.find('img')
     this.txtViewer = getElementWithTestId('viewer-text')
     this.txtViewerContent = this.txtViewer.find(
       '[class*="pho-viewer-textviewer-content"]'
@@ -46,14 +27,6 @@ export default class Page {
     this.btnPdfViewerDownload = this.pdfViewer.find('#download')
   }
 
-  //@param { bool } isSingleShareFile : set to true only when checking viewer on a single shared file
-  async waitForLoading() {
-    await t.expect(this.spinner.exists).notOk('Spinner still spinning')
-    await isExistingAndVisibile(this.viewerWrapper, 'Viewer Wrapper')
-    await isExistingAndVisibile(this.viewerControls, 'Viewer Controls')
-    console.log('Viewer Ok')
-  }
-
   //@param {String} fileName
   async openViewerForFile(fileName) {
     await t
@@ -63,45 +36,6 @@ export default class Page {
 
     await this.waitForLoading()
     console.log(`Navigation to ${fileName} OK!`)
-  }
-
-  //@param { bool } exitWithEsc : true to exit by pressing esc, false to click on the button
-  async closeViewer(exitWithEsc) {
-    await t.hover(this.viewerWrapper)
-    await isExistingAndVisibile(this.viewerBtnClose, 'Close button')
-    exitWithEsc ? await t.pressKey('esc') : await t.click(this.viewerBtnClose)
-  }
-
-  //@param {number} index: index of open file (need to know if it's first or last file)
-  async navigateToNextFile(index) {
-    if (index == t.ctx.totalFilesCount - 1) {
-      //this is the last picture, so next button does not exist
-      await t
-        .expect(this.viewerNavNext.exists)
-        .notOk('Next button on last picture')
-    } else {
-      await t
-        .hover(this.viewerNavNext) //not last photo, so next button should exists
-        .expect(this.viewerNavNextBtn.visible)
-        .ok('Next arrow does not show up')
-        .click(this.viewerNavNextBtn)
-    }
-  }
-
-  //@param {number} index: index of open file (need to know if it's first or last file)
-  async navigateToPrevFile(index) {
-    if (index == 0) {
-      //this is the 1st picture, so previous button does not exist
-      await t
-        .expect(this.viewerNavPrevious.exists)
-        .notOk('Previous button on first picture')
-    } else {
-      await t
-        .hover(this.viewerNavPrevious) //not 1st photo, so previous button should exists
-        .expect(this.viewerNavPreviousBtn.visible)
-        .ok('Previous arrow does not show up')
-        .click(this.viewerNavPrevious)
-    }
   }
 
   //@param {String} screenshotPath : path for screenshots taken in this test
@@ -158,16 +92,9 @@ export default class Page {
   }
 
   //download using the common download button
-  async checkCommonViewerDownload(fileName) {
+  async openFileAndCheckCommonViewerDownload(fileName) {
     await this.openViewerForFile(fileName)
-    await t.hover(this.viewerWrapper)
-    await isExistingAndVisibile(
-      this.btnDownloadViewerToolbar,
-      'Download button in toolbar'
-    )
-    await t
-      .setNativeDialogHandler(() => true)
-      .click(this.btnDownloadViewerToolbar)
+    await this.checkCommonViewerDownload()
     await this.closeViewer({
       exitWithEsc: true
     })
@@ -203,15 +130,6 @@ export default class Page {
   async checkTextViewer() {
     await isExistingAndVisibile(this.txtViewer, 'text viewer')
     await isExistingAndVisibile(this.txtViewerContent, 'text viewer controls')
-  }
-
-  //Specific check for imageViewer
-  async checkImageViewer() {
-    await isExistingAndVisibile(this.imageViewer, 'image viewer')
-    await isExistingAndVisibile(
-      this.imageViewerContent,
-      'image viewer controls'
-    )
   }
 
   //Specific check for no viewer : other download btn
