+++ testcafe/tests/photos/create_empty_album_scenario.js
@@ -1,33 +1,33 @@
 import { photosUser } from '../helpers/roles'
 import { TESTCAFE_PHOTOS_URL } from '../helpers/utils'
 import { ALBUM_DATE_TIME } from '../helpers/data'
-import PhotoPage from '../pages/photos-model'
+import Timeline from '../pages/photos/photos-timeline-model'
 import AlbumPage from '../pages/photos-album-model'
 
-const photoPage = new PhotoPage()
+const timelinePage = new Timeline()
 const photoAlbumPage = new AlbumPage()
 
 fixture`Create new empty album and add photos`
   .page`${TESTCAFE_PHOTOS_URL}/`.beforeEach(async t => {
   await t.useRole(photosUser)
-  await photoPage.waitForLoading()
-  await photoPage.initPhotosCount()
+  await timelinePage.waitForLoading()
+  await timelinePage.initPhotosCount()
 })
 
 test('Go into Album view, and check that there is no album', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.checkEmptyAlbum()
 })
 
 test('Go into Album view, and create new empty album', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.addNewAlbum(ALBUM_DATE_TIME, 0)
   await photoAlbumPage.checkAlbumPage(ALBUM_DATE_TIME, 0)
   //we need to check the album page, just after the redirection from album creation, hence this step being in this test
 })
 
 test('Go to ALBUM_DATE_TIME, and rename it (exit by pressing "enter")', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.goToAlbum(ALBUM_DATE_TIME)
   await photoAlbumPage.checkAlbumPage(ALBUM_DATE_TIME, 0)
   await photoAlbumPage.renameAlbum(ALBUM_DATE_TIME, `New_${ALBUM_DATE_TIME}`, {
@@ -37,7 +37,7 @@ test('Go to ALBUM_DATE_TIME, and rename it (exit by pressing "enter")', async ()
 })
 
 test('Go to New_ALBUM_DATE_TIME, and rename it (exit by clicking away)', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.goToAlbum(`New_${ALBUM_DATE_TIME}`)
   await photoAlbumPage.checkAlbumPage(`New_${ALBUM_DATE_TIME}`, 0)
   await photoAlbumPage.renameAlbum(
@@ -49,7 +49,7 @@ test('Go to New_ALBUM_DATE_TIME, and rename it (exit by clicking away)', async (
 })
 
 test('Go to New2_ALBUM_DATE_TIME, and add 2 more photos', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.goToAlbum(`New2_${ALBUM_DATE_TIME}`)
   await photoAlbumPage.addPhotosToAlbum(`New2_${ALBUM_DATE_TIME}`, 0, 2)
   await photoAlbumPage.backToAlbumsList()
@@ -57,13 +57,13 @@ test('Go to New2_ALBUM_DATE_TIME, and add 2 more photos', async () => {
 })
 
 test('Go to New2_ALBUM_DATE_TIME, and remove the 1st photos', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.goToAlbum(`New2_${ALBUM_DATE_TIME}`)
   await photoAlbumPage.removePhoto(1)
 })
 
 test('Go to New2_ALBUM_DATE_TIME, and delete it', async () => {
-  await photoPage.goToAlbums()
+  await timelinePage.goToAlbums()
   await photoAlbumPage.goToAlbum(`New2_${ALBUM_DATE_TIME}`)
   await photoAlbumPage.deleteAlbum()
   await photoAlbumPage.waitForLoading()
