+++ testcafe/tests/photos/photos_crud.js
@@ -1,49 +1,57 @@
 import { photosUser } from '../helpers/roles' //import roles for login
 import { TESTCAFE_PHOTOS_URL } from '../helpers/utils'
 import random from 'lodash/random'
-import Page from '../pages/photos-model'
+import Timeline from '../pages/photos/photos-timeline-model'
+import Viewer from '../pages/photos-viewer/photos-viewer-model'
 
-const page = new Page()
+const timelinePage = new Timeline()
+const photoViewer = new Viewer()
 
 fixture`PHOTOS - CRUD`.page`${TESTCAFE_PHOTOS_URL}/`.beforeEach(async t => {
   await t.useRole(photosUser)
-  await page.waitForLoading()
-  await page.initPhotosCount()
+  await timelinePage.waitForLoading()
+  await timelinePage.initPhotosCount()
 })
 
 test('Select 1 pic from Photos view', async () => {
   //Selection bar shows up. It includes AddtoAlbun, Download and Delete buttons
-  await page.selectPhotos(1)
-  await page.checkPhotobar()
+  await timelinePage.selectPhotos(1)
+  await timelinePage.checkPhotobar()
 })
 
 test('Select 3 pic from Photos view', async () => {
   //Selection bar shows up. It includes AddtoAlbun, Download and Delete buttons
-  await page.selectPhotos(3)
-  await page.checkPhotobar()
+  await timelinePage.selectPhotos(3)
+  await timelinePage.checkPhotobar()
 })
 
 test('Open 1st pic', async () => {
   //Right arrow shows up. Navigatio to other pics is OK, Closing pic (X or 'esc') is Ok
-  await page.openPhotoFullscreen(0)
-  await page.navigateToNextPhoto(0)
-  await page.closePhotoFullscreenX()
-
-  await page.openPhotoFullscreen(0)
-  await page.navigateToPrevPhoto(0)
-  await page.closePhotoFullscreenEsc()
+  await photoViewer.openPhotoFullscreen(0)
+  await photoViewer.navigateToNextPhoto(0)
+  await photoViewer.closeViewer({
+    exitWithEsc: true
+  })
+  await photoViewer.openPhotoFullscreen(0)
+  await photoViewer.navigateToPrevPhoto(0)
+  await photoViewer.closeViewer({
+    exitWithEsc: false
+  })
 })
 
 test('Open Last pic', async t => {
   //Left arrow shows up. Navigatio to other pics is OK, Closing pic (X or 'esc') is Ok
 
-  await page.openPhotoFullscreen(t.ctx.allPhotosStartCount - 1)
-  await page.navigateToNextPhoto(t.ctx.allPhotosStartCount - 1)
-  await page.closePhotoFullscreenX()
-
-  await page.openPhotoFullscreen(t.ctx.allPhotosStartCount - 1)
-  await page.navigateToPrevPhoto(t.ctx.allPhotosStartCount - 1)
-  await page.closePhotoFullscreenEsc()
+  await photoViewer.openPhotoFullscreen(t.ctx.allPhotosStartCount - 1)
+  await photoViewer.navigateToNextPhoto(t.ctx.allPhotosStartCount - 1)
+  await photoViewer.closeViewer({
+    exitWithEsc: true
+  })
+  await photoViewer.openPhotoFullscreen(t.ctx.allPhotosStartCount - 1)
+  await photoViewer.navigateToPrevPhoto(t.ctx.allPhotosStartCount - 1)
+  await photoViewer.closeViewer({
+    exitWithEsc: false
+  })
 })
 
 test('Open a random pic (not first nor last)', async t => {
@@ -53,10 +61,14 @@ test('Open a random pic (not first nor last)', async t => {
   const photoIndex = random(1, t.ctx.allPhotosStartCount - 2)
 
   console.log('Open random pic  > photoIndex ' + photoIndex)
-  await page.openPhotoFullscreen(photoIndex)
-  await page.navigateToNextPhoto(photoIndex)
-  await page.closePhotoFullscreenX()
-
-  await page.openPhotoFullscreen(photoIndex)
-  await page.navigateToPrevPhoto(photoIndex)
+  await photoViewer.openPhotoFullscreen(photoIndex)
+  await photoViewer.navigateToNextPhoto(photoIndex)
+  await photoViewer.closeViewer({
+    exitWithEsc: true
+  })
+  await photoViewer.openPhotoFullscreen(photoIndex)
+  await photoViewer.navigateToPrevPhoto(photoIndex)
+  await photoViewer.closeViewer({
+    exitWithEsc: false
+  })
 })
